requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
pydantic==1.10.9
pytest==7.4.2
python-dotenv==1.0.0
